@@ -3,6 +3,7 @@
   "private": true,
   "scripts": {
     "start": "yarn docs:dev",
+    "benchmark:browser": "yarn workspace benchmark browser",
     "docs:dev": "yarn workspace docs dev",
     "docs:api": "yarn docs:api:build",
     "docs:api:build": "cross-env BABEL_ENV=development babel-node -x .ts,.tsx,.js ./docs/scripts/buildApi.ts ./docs/pages/api-docs/data-grid",
@@ -61,25 +62,25 @@
     "@emotion/styled": "^11.3.0",
     "@eps1lon/enzyme-adapter-react-17": "^0.1.0",
     "@material-ui/monorepo": "https://github.com/mui-org/material-ui.git#master",
-    "@mui/icons-material": "^5.0.0",
-    "@mui/material": "^5.0.0",
-    "@mui/styles": "^5.0.0",
-    "@mui/utils": "^5.0.0",
+    "@mui/icons-material": "^5.0.1",
+    "@mui/material": "^5.0.1",
+    "@mui/styles": "^5.0.1",
+    "@mui/utils": "^5.0.1",
     "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-node-resolve": "^13.0.4",
+    "@rollup/plugin-node-resolve": "^13.0.5",
     "@rollup/plugin-replace": "^3.0.0",
     "@testing-library/react": "^12.1.0",
-    "@types/chai": "^4.2.21",
+    "@types/chai": "^4.2.22",
     "@types/chai-dom": "^0.0.11",
     "@types/enzyme": "^3.10.9",
     "@types/mocha": "^9.0.0",
-    "@types/node": "^16.9.4",
-    "@types/prettier": "^2.3.2",
-    "@types/react": "^17.0.22",
+    "@types/node": "^16.10.1",
+    "@types/prettier": "^2.4.0",
+    "@types/react": "^17.0.24",
     "@types/react-dom": "^17.0.9",
     "@types/requestidlecallback": "^0.3.4",
-    "@types/sinon": "^10.0.2",
-    "@types/yargs": "^17.0.0",
+    "@types/sinon": "^10.0.4",
+    "@types/yargs": "^17.0.3",
     "@typescript-eslint/eslint-plugin": "^4.31.2",
     "@typescript-eslint/parser": "^4.31.2",
     "argos-cli": "^0.3.3",
@@ -102,7 +103,7 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-mocha": "^9.0.0",
     "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-react": "^7.25.3",
+    "eslint-plugin-react": "^7.26.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "format-util": "^1.0.5",
     "glob-gitignore": "^1.0.14",
@@ -112,15 +113,15 @@
     "karma-chrome-launcher": "^3.1.0",
     "karma-mocha": "^2.0.1",
     "karma-sourcemap-loader": "^0.3.8",
-    "karma-webpack": "^4.0.2",
+    "karma-webpack": "^5.0.0",
     "lerna": "^4.0.0",
-    "mocha": "^9.1.1",
+    "mocha": "^9.1.2",
     "nyc": "^15.1.0",
-    "playwright": "^1.14.1",
+    "playwright": "^1.15.0",
     "prettier": "^2.4.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "rollup": "^2.56.3",
+    "rollup": "^2.57.0",
     "rollup-plugin-cleaner": "^1.0.0",
     "rollup-plugin-command": "^1.1.3",
     "rollup-plugin-commonjs": "^10.1.0",
@@ -136,9 +137,9 @@
     "stylelint-processor-styled-components": "^1.10.0",
     "typedoc": "^0.22.4",
     "typescript": "^4.4.3",
-    "webpack": "^4.46.0",
+    "webpack": "^5.54.0",
     "webpack-cli": "^4.8.0",
-    "yargs": "^17.0.1",
+    "yargs": "^17.2.1",
     "yarn-deduplicate": "^3.1.0"
   },
   "setupFiles": [
@@ -146,6 +147,7 @@
   ],
   "workspaces": {
     "packages": [
+      "benchmark",
       "packages/*",
       "!packages/grid",
       "packages/grid/*",
@@ -154,5 +156,8 @@
     "nohoist": [
       "**/@material-ui/monorepo"
     ]
+  },
+  "dependencies": {
+    "html-webpack-plugin": "^5.3.2"
   }
 }
