@@ -1,3 +1,5 @@
+const path = require('path');
+
 let defaultPresets;
 
 // We release a ES version of Material-UI.
@@ -17,12 +19,17 @@ if (process.env.BABEL_ENV === 'es') {
   ];
 }
 
+function resolveAliasPath(relativeToBabelConf) {
+  const resolvedPath = path.relative(process.cwd(), path.resolve(__dirname, relativeToBabelConf));
+  return `./${resolvedPath.replace('\\', '/')}`;
+}
+
 const defaultAlias = {
-  '@mui/x-data-grid': './packages/grid/data-grid/src',
-  '@mui/x-data-grid-generator': './packages/grid/x-grid-data-generator/src',
-  '@mui/x-data-grid-pro': './packages/grid/x-grid/src',
-  '@mui/x-license-pro': './packages/x-license/src',
-  docs: './node_modules/@material-ui/monorepo/docs',
+  '@mui/x-data-grid': resolveAliasPath('./packages/grid/data-grid/src'),
+  '@mui/x-data-grid-generator': resolveAliasPath('./packages/grid/x-grid-data-generator/src'),
+  '@mui/x-data-grid-pro': resolveAliasPath('./packages/grid/x-grid/src'),
+  '@mui/x-license-pro': resolveAliasPath('./packages/x-license/src'),
+  docs: resolveAliasPath('./node_modules/@material-ui/monorepo/docs'),
 };
 
 module.exports = {
@@ -55,6 +62,16 @@ module.exports = {
     coverage: {},
     test: {
       sourceMaps: 'both',
+    },
+    benchmark: {
+      plugins: [
+        [
+          'babel-plugin-module-resolver',
+          {
+            alias: defaultAlias,
+          },
+        ],
+      ],
     },
   },
 };
