@@ -22,13 +22,13 @@ import {
   randomUpdatedDate,
   randomId,
 } from '@mui/x-data-grid-generator';
-import { createTheme, Theme } from '@mui/material/styles';
+import { createTheme } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
 
 const defaultTheme = createTheme();
 
 const useStyles = makeStyles(
-  (theme: Theme) => ({
+  (theme) => ({
     actions: {
       color: theme.palette.text.secondary,
     },
