@@ -8,16 +8,11 @@ import {
   randomUpdatedDate,
 } from '@mui/x-data-grid-generator';
 
-// TODO v5: remove
-function getThemePaletteMode(palette) {
-  return palette.type || palette.mode;
-}
-
 const defaultTheme = createTheme();
 const useStyles = makeStyles(
   (theme) => {
     const backgroundColor =
-      getThemePaletteMode(theme.palette) === 'dark' ? '#376331' : 'rgb(217 243 190)';
+      theme.palette.mode === 'dark' ? '#376331' : 'rgb(217 243 190)';
     return {
       root: {
         '& .MuiDataGrid-cell--editable': {
