@@ -3,15 +3,10 @@ import { createTheme } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
 import { useGridApiRef, DataGridPro } from '@mui/x-data-grid-pro';
 
-// TODO v5: remove
-function getThemePaletteMode(palette) {
-  return palette.type || palette.mode;
-}
-
 const defaultTheme = createTheme();
 const useStyles = makeStyles(
   (theme) => {
-    const isDark = getThemePaletteMode(theme.palette) === 'dark';
+    const isDark = theme.palette.mode === 'dark';
 
     return {
       root: {
