@@ -4,23 +4,14 @@ import { useDemoData } from '@mui/x-data-grid-generator';
 import { createTheme, darken, lighten } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
 
-// TODO v5: remove
-function getThemePaletteMode(palette) {
-  return palette.type || palette.mode;
-}
-
 const defaultTheme = createTheme();
 const useStyles = makeStyles(
   (theme) => {
     const getBackgroundColor = (color) =>
-      getThemePaletteMode(theme.palette) === 'dark'
-        ? darken(color, 0.6)
-        : lighten(color, 0.6);
+      theme.palette.mode === 'dark' ? darken(color, 0.6) : lighten(color, 0.6);
 
     const getHoverBackgroundColor = (color) =>
-      getThemePaletteMode(theme.palette) === 'dark'
-        ? darken(color, 0.5)
-        : lighten(color, 0.5);
+      theme.palette.mode === 'dark' ? darken(color, 0.5) : lighten(color, 0.5);
 
     return {
       root: {
