@@ -1,17 +1,16 @@
 import * as React from 'react';
-import TablePagination from '@mui/material/TablePagination';
-import { Theme } from '@mui/material/styles';
+import TablePagination, { TablePaginationProps } from '@mui/material/TablePagination';
+import { createTheme } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
 import { useGridSelector } from '../hooks/features/core/useGridSelector';
 import { gridPaginationSelector } from '../hooks/features/pagination/gridPaginationSelector';
 import { useGridApiContext } from '../hooks/root/useGridApiContext';
-import { createTheme, getMuiVersion } from '../utils/utils';
 import { useGridRootProps } from '../hooks/utils/useGridRootProps';
 
 const defaultTheme = createTheme();
 // Used to hide the Rows per page selector on small devices
 const useStyles = makeStyles(
-  (theme: Theme) => ({
+  (theme) => ({
     selectLabel: {
       display: 'none',
       [theme.breakpoints.up('sm')]: {
@@ -59,26 +58,12 @@ export const GridPagination = React.forwardRef<
     [apiRef],
   );
 
-  const handlePageChange = React.useCallback(
-    (event: any, page: number) => {
+  const handlePageChange = React.useCallback<TablePaginationProps['onPageChange']>(
+    (event, page) => {
       apiRef.current.setPage(page);
     },
     [apiRef],
   );
-
-  const getPaginationChangeHandlers = () => {
-    if (getMuiVersion() !== 'v4') {
-      return {
-        onPageChange: handlePageChange,
-        onRowsPerPageChange: handlePageSizeChange,
-      };
-    }
-
-    return {
-      onChangePage: handlePageChange,
-      onChangeRowsPerPage: handlePageSizeChange,
-    };
-  };
 
   if (process.env.NODE_ENV !== 'production') {
     // eslint-disable-next-line react-hooks/rules-of-hooks
@@ -102,13 +87,10 @@ export const GridPagination = React.forwardRef<
   }
 
   return (
-    // @ts-ignore TODO remove once upgraded v4 support is dropped
     <TablePagination
       ref={ref}
       classes={{
-        ...(getMuiVersion() === 'v5'
-          ? { selectLabel: classes.selectLabel }
-          : { caption: classes.caption }),
+        selectLabel: classes.selectLabel,
         input: classes.input,
       }}
       component="div"
@@ -120,8 +102,9 @@ export const GridPagination = React.forwardRef<
           : []
       }
       rowsPerPage={paginationState.pageSize}
+      onPageChange={handlePageChange}
+      onRowsPerPageChange={handlePageSizeChange}
       {...apiRef.current.getLocaleText('MuiTablePagination')}
-      {...getPaginationChangeHandlers()}
       {...props}
     />
   );
