@@ -1,7 +1,5 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { gridColumnReorderDragColSelector } from '../../hooks/features/columnReorder/columnReorderSelector';
-import { gridResizingColumnFieldSelector } from '../../hooks/features/columnResize/columnResizeSelector';
 import { useGridSelector } from '../../hooks/features/core/useGridSelector';
 import { filterGridColumnLookupSelector } from '../../hooks/features/filter/gridFilterSelector';
 import {
@@ -21,15 +19,15 @@ import { gridScrollBarSizeSelector } from '../../hooks/root/gridContainerSizesSe
 
 export interface GridColumnHeadersItemCollectionProps {
   columns: GridStateColDef[];
+  dragCol: string;
+  resizeCol: string;
 }
 
 function GridColumnHeadersItemCollection(props: GridColumnHeadersItemCollectionProps) {
-  const { columns } = props;
+  const { columns, dragCol, resizeCol } = props;
   const apiRef = useGridApiContext();
   const sortColumnLookup = useGridSelector(apiRef, gridSortColumnLookupSelector);
   const filterColumnLookup = useGridSelector(apiRef, filterGridColumnLookupSelector);
-  const dragCol = useGridSelector(apiRef, gridColumnReorderDragColSelector);
-  const resizingColumnField = useGridSelector(apiRef, gridResizingColumnFieldSelector);
   const columnHeaderFocus = useGridSelector(apiRef, gridFocusColumnHeaderSelector);
   const renderCtx = useGridSelector(apiRef, gridRenderingSelector).renderContext;
   const tabIndexState = useGridSelector(apiRef, gridTabIndexColumnHeaderSelector);
@@ -69,7 +67,7 @@ function GridColumnHeadersItemCollection(props: GridColumnHeadersItemCollectionP
         isDragging={col.field === dragCol}
         column={col}
         colIndex={colIndex}
-        isResizing={resizingColumnField === col.field}
+        isResizing={resizeCol === col.field}
         isLastColumn={colIndex === columns.length - 1}
         extendRowFullWidth={!rootProps.disableExtendRowFullWidth}
         hasScrollX={scrollBarState.hasScrollX}
@@ -89,6 +87,8 @@ GridColumnHeadersItemCollection.propTypes = {
   // | To update them edit the TypeScript types and run "yarn proptypes"  |
   // ----------------------------------------------------------------------
   columns: PropTypes.arrayOf(PropTypes.object).isRequired,
+  dragCol: PropTypes.string.isRequired,
+  resizeCol: PropTypes.string.isRequired,
 } as any;
 
 export { GridColumnHeadersItemCollection };
