@@ -1,13 +1,12 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
-import { Theme } from '@mui/material/styles';
+import { createTheme } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
 import ClickAwayListener from '@mui/material/ClickAwayListener';
 import Paper from '@mui/material/Paper';
 import Popper, { PopperProps } from '@mui/material/Popper';
 import { useGridApiContext } from '../../hooks/root/useGridApiContext';
-import { getMuiVersion, createTheme } from '../../utils/utils';
 import { isEscapeKey } from '../../utils/keyboardUtils';
 import {
   InternalStandardProps as StandardProps,
@@ -32,7 +31,7 @@ export interface GridPanelProps extends StandardProps<PopperProps, 'children'> {
 
 const defaultTheme = createTheme();
 const useStyles = makeStyles(
-  (theme: Theme) => ({
+  (theme) => ({
     root: {
       zIndex: theme.zIndex.modal,
     },
@@ -52,23 +51,6 @@ const GridPanel = React.forwardRef<HTMLDivElement, GridPanelProps>((props, ref) 
   const { children, className, open, classes: classesProp, ...other } = props;
   const classes = useStyles(props);
   const apiRef = useGridApiContext();
-
-  const getPopperModifiers = (): any => {
-    if (getMuiVersion() === 'v5') {
-      return [
-        {
-          name: 'flip',
-          enabled: false,
-        },
-      ];
-    }
-
-    return {
-      flip: {
-        enabled: false,
-      },
-    };
-  };
 
   const handleClickAway = React.useCallback(() => {
     apiRef.current.hidePreferences();
@@ -96,7 +78,12 @@ const GridPanel = React.forwardRef<HTMLDivElement, GridPanelProps>((props, ref) 
       className={clsx(className, classes.root)}
       open={open}
       anchorEl={anchorEl}
-      modifiers={getPopperModifiers()}
+      modifiers={[
+        {
+          name: 'flip',
+          enabled: false,
+        },
+      ]}
       {...other}
     >
       <ClickAwayListener onClickAway={handleClickAway}>
