@@ -1,12 +1,11 @@
 import * as React from 'react';
 import clsx from 'clsx';
-import { Theme } from '@mui/material/styles';
+import { createTheme } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
-import { createTheme } from '../../utils/utils';
 
 const defaultTheme = createTheme();
 const useStyles = makeStyles(
-  (theme: Theme) => ({
+  (theme) => ({
     root: {
       padding: theme.spacing(1),
     },
