@@ -1,6 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { unstable_useId as useId } from '@mui/material/utils';
+import { unstable_useId as useId, useForkRef } from '@mui/material/utils';
 import MenuList from '@mui/material/MenuList';
 import Button, { ButtonProps } from '@mui/material/Button';
 import MenuItem from '@mui/material/MenuItem';
@@ -32,7 +32,10 @@ const GridToolbarExport = React.forwardRef<HTMLButtonElement, GridToolbarExportP
     const rootProps = useGridRootProps();
     const buttonId = useId();
     const menuId = useId();
-    const [anchorEl, setAnchorEl] = React.useState(null);
+
+    const [open, setOpen] = React.useState(false);
+    const buttonRef = React.useRef<HTMLButtonElement>(null);
+    const handleRef = useForkRef(ref, buttonRef);
 
     const exportOptions: Array<GridExportOption> = [];
     exportOptions.push({
@@ -42,16 +45,16 @@ const GridToolbarExport = React.forwardRef<HTMLButtonElement, GridToolbarExportP
     });
 
     const handleMenuOpen = (event) => {
-      setAnchorEl(event.currentTarget);
+      setOpen(true);
       onClick?.(event);
     };
-    const handleMenuClose = () => setAnchorEl(null);
+    const handleMenuClose = () => setOpen(false);
     const handleExport = (option: GridExportOption) => () => {
       if (option.format === 'csv') {
         apiRef.current.exportDataAsCsv(option.formatOptions);
       }
 
-      setAnchorEl(null);
+      setOpen(false);
     };
 
     const handleListKeyDown = (event: React.KeyboardEvent) => {
@@ -66,11 +69,11 @@ const GridToolbarExport = React.forwardRef<HTMLButtonElement, GridToolbarExportP
     return (
       <React.Fragment>
         <Button
-          ref={ref}
+          ref={handleRef}
           color="primary"
           size="small"
           startIcon={<rootProps.components.ExportIcon />}
-          aria-expanded={anchorEl ? 'true' : undefined}
+          aria-expanded={open ? 'true' : undefined}
           aria-label={apiRef.current.getLocaleText('toolbarExportLabel')}
           aria-haspopup="menu"
           aria-labelledby={menuId}
@@ -81,8 +84,8 @@ const GridToolbarExport = React.forwardRef<HTMLButtonElement, GridToolbarExportP
           {apiRef.current.getLocaleText('toolbarExport')}
         </Button>
         <GridMenu
-          open={Boolean(anchorEl)}
-          target={anchorEl}
+          open={open}
+          target={buttonRef.current}
           onClickAway={handleMenuClose}
           position="bottom-start"
         >
@@ -91,7 +94,7 @@ const GridToolbarExport = React.forwardRef<HTMLButtonElement, GridToolbarExportP
             className="MuiDataGrid-gridMenuList"
             aria-labelledby={buttonId}
             onKeyDown={handleListKeyDown}
-            autoFocusItem={Boolean(anchorEl)}
+            autoFocusItem={open}
           >
             {exportOptions.map((option, index) => (
               <MenuItem key={index} onClick={handleExport(option)}>
