@@ -1,6 +1,5 @@
 import { GridLocaleText } from '../models/api/gridLocaleTextApi';
 import { GridMergedOptions } from '../models/gridOptions';
-import { getMuiVersion } from './utils';
 
 interface LocalizationV4 {
   props: {
@@ -21,31 +20,15 @@ export type Localization = LocalizationV4 | LocalizationV5;
 export const getGridLocalization = (
   gridTranslations: Partial<GridLocaleText>,
   coreTranslations?,
-): Localization => {
-  if (getMuiVersion() === 'v5') {
-    return {
-      components: {
-        MuiDataGrid: {
-          defaultProps: {
-            localeText: {
-              ...gridTranslations,
-              MuiTablePagination:
-                coreTranslations?.components?.MuiTablePagination.defaultProps || {},
-            },
-          },
-        },
-      },
-    };
-  }
-
-  return {
-    props: {
-      MuiDataGrid: {
+): Localization => ({
+  components: {
+    MuiDataGrid: {
+      defaultProps: {
         localeText: {
           ...gridTranslations,
-          MuiTablePagination: coreTranslations?.props?.MuiTablePagination || {},
+          MuiTablePagination: coreTranslations?.components?.MuiTablePagination.defaultProps || {},
         },
       },
     },
-  };
-};
+  },
+});
