@@ -1,13 +1,12 @@
 import * as React from 'react';
 import clsx from 'clsx';
-import { Theme } from '@mui/material/styles';
+import { createTheme } from '@mui/material/styles';
 import { createStyles, makeStyles } from '@mui/styles';
 import { GridCellParams } from '../../../_modules_/grid';
-import { createTheme, getThemePaletteMode } from '../../../_modules_/grid/utils/utils';
 
 const defaultTheme = createTheme();
 const useStyles = makeStyles(
-  (theme: Theme) =>
+  (theme) =>
     createStyles({
       root: {
         width: '100%',
@@ -15,15 +14,11 @@ const useStyles = makeStyles(
       },
       positive: {
         color:
-          getThemePaletteMode(theme.palette) === 'light'
-            ? theme.palette.success.dark
-            : theme.palette.success.light,
+          theme.palette.mode === 'light' ? theme.palette.success.dark : theme.palette.success.light,
       },
       negative: {
         color:
-          getThemePaletteMode(theme.palette) === 'light'
-            ? theme.palette.error.dark
-            : theme.palette.error.light,
+          theme.palette.mode === 'light' ? theme.palette.error.dark : theme.palette.error.light,
       },
     }),
   { defaultTheme },
