@@ -4,5 +4,5 @@ export default create({
   base: 'light',
   brandTitle: 'Material-UI X Grid storybook',
   brandUrl: 'https://material-ui.com/',
-  brandImage: 'https://material-ui.com/static/logo_raw.svg',
+  brandImage: 'https://material-ui.com/static/logo_raw.svg', // TODO replace with a smaller logo
 });
